+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Buffer with no underlying file now use `&` for the buffer status.
 - While `flycheck` is running, the buffer name is shown underlined but
   shadowed.
+- When there's not enough space to show both sides, only the left side
+  is displayed, indicating the elision with an ellipsis.
 
 ### Fixed